Рефакторинг, вывод хуманизированных размеров
