+++ less8/go.mod
@@ -2,4 +2,7 @@ module github.com/vsurkov/go-gb2
 
 go 1.18
 
-require github.com/namsral/flag v1.7.4-pre
+require (
+	github.com/dustin/go-humanize v1.0.0
+	github.com/namsral/flag v1.7.4-pre
+)
